outsource helper methods
