+++ src/shared/src/helper/game/index.ts
@@ -1,0 +1,3 @@
+import { calculateScore } from './gameHelper';
+
+export { calculateScore };
