+++ src/shared/src/helper/pokeApi/index.ts
@@ -1,0 +1,3 @@
+import { fetchPokemonGenerations } from './pokeApiHelper';
+
+export { fetchPokemonGenerations };
